backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.5
langgraph==0.0.20
pillow>=10.0.0
PyPDF2==3.0.1
aiofiles==23.2.1
